fixedd bug of countries fetching only once
