+++ src/Redux/countries/CountriesReducer.js
@@ -1,14 +1,25 @@
 import { fetchCountries } from '../../API/CovidAPI';
+import { formattedDate } from '../../Logic/DateFormatter';
 
 const GET_COUNTRIES = 'Capstone-3/countries/GET_COUNTRIES';
 
 const initialState = [];
 
 const getCountries = () => async (dispatch) => {
-  const result = await fetchCountries();
+  const data = await fetchCountries();
+  const countriesList = data.dates[formattedDate].countries;
+  const mappedList = Object.keys(countriesList).map((key) => ({
+    id: key,
+    name: countriesList[key].name,
+    today_confirmed: countriesList[key].today_confirmed,
+    today_deaths: countriesList[key].today_deaths,
+    today_recovered: countriesList[key].today_recovered,
+    source: countriesList[key].source,
+    today_open_cases: countriesList[key].today_open_cases,
+  }));
   dispatch({
     type: GET_COUNTRIES,
-    payload: result,
+    payload: mappedList,
   });
 };
